requirements.txt
gradio==5.9.0
langchain_openai==0.2.12
langchain_community==0.3.12
langchain-core==0.3.25
pinecone-client==5.0.1
pymongo==4.10.1
python-dotenv==1.0.1
unidecode==1.3.8
unstructured==0.16.11
